Refactor: Enhance Promote Tools & Streamline Seller Hub
This commit introduces a major enhancement to the "Promote Tools" section by splitting it into "Overview" and "Promoted Shows" tabs. It also includes significant UI/UX improvements across the Seller Hub, shipping profile management, and live streaming interface.

**Key Changes:**

*   **Promote Tools Overhaul:**
    *   The "Promote Tools" screen is now tabbed, separating analytics from the list of promoted shows.
    *   **New "Overview" Tab:** A new `OverviewFragment` displays detailed promotion metrics like "Audience Reached," "Discovery Impact," and "Buyers Converted," complete with pro tips for sellers.
    *   **New "Promoted Shows" Tab:** A dedicated `PromotedShowsFragment` lists all shows marked for promotion, allowing sellers to view and manage them.
    *   A `ViewPager2` is now used to manage the tab content, improving navigation and structure.

*   **Shipping & Inventory Enhancements:**
    *   **Shipping Profiles:**
        *   Added "Bundle in one package" (`maxItems`) and "Pay for additional item weight" (`additionalWeight`) as boolean options when creating a shipping profile.
        *   The shipping profile list now displays these new bundling and weight options.
    *   **Free Pickup:** The "Free Pickup" setting is now fetched and displayed correctly, showing its "ON/OFF" status in the shipping options list.
    *   **Inventory UI:** The filter icon (`ic_product_filter`) in the inventory screen is now tinted with the primary color for better visibility.

*   **UI/UX Improvements:**
    *   **Search Bar Icons:** The search icon tint in `ExploreFragment`, `HomeFragment`, `MyOrdersFragment`, and `InventoryFragment` has been changed to `outlineVariant` for better UI consistency.
    *   **Explore Screen:** The category filter chips are reordered to show "All," "Recommended," and then "Popular" for a more logical flow.
    *   **My Orders:** Fixed a bug where the order list wouldn't refresh correctly when changing the status filter (e.g., "Processing," "Completed").
    *   **Live Stream (`AgoraPublisherActivity`):** Refactored the layout to use padding for handling system window insets, providing a cleaner look and removing manual margin adjustments.
    *   **Image Loading:** Fixed an issue where product images in the live stream (`AgoraPublisherActivity`, `WatchStreamFragment`) would not load correctly if the URL was missing the base URL.
    *   **Pro Tip Styling:** The "Pro Tip" item layout (`item_pro_tip.xml`) has been updated with a `primaryContainer` background and a primary-colored icon for better visual distinction.
    *   Minor UI tweaks in `AffiliateProgramFragment`, `my_orders_item.xml`, and `NotificationActivity` for improved text display and layout.

*   **Code Cleanup:**
    *   Removed several unused and redundant `strings.xml` files for different configurations (`-land`, `-night`, `-v23`, etc.), cleaning up project resources.
    *   Removed an obsolete `RecyclerView` (`gridRecycler`) from the `seller_hub_view.xml` layout.

+++ app/src/main/java/io/bidswipe/app/network/response/GetShowOverviewResponse.kt
@@ -17,7 +17,7 @@ data class GetShowOverviewResponse(
     @Keep
     data class Data(
         @SerializedName("contributions_count")
-        val contributionsCount: Int?,
+        val contributionsCount: Double?,
         @SerializedName("new_followers")
         val newFollowers: Int?,
         @SerializedName("file_url")
